Always run installs to ensure dependencies are installed
I accidentally deployed a copy of www to production that did not
include a query plugin needed by the new technology page. This change
is a little ham-fisted, but ensures dependencies are always installed
locally by running bundle, npm, and bower installs as part of the
setup, server, and deploy tasks.