Add a support topic for private images
Instead of only having the blog post as documentation, add private
images as an actual support topic as well. This also fixes two glaring
omissions/inconsistencies with the documentation, vis-a-vis with what
the code actually does:

1) If the registry host is not docker.io (most notably images hosted
on quay.io for example), APTIBLE_DOCKER_IMAGE needs the registry
host specified in the image name; so instead of my_repo/my_image a user
needs to specify quay.io/my_repo/my_image.
2) If APTIBLE_DOCKER_IMAGE is specified, and if there's a Dockerfile
in the same app repo, it will be totally ignored. If you want to extend
a private image, you need to omit APTIBLE_DOCKER_IMAGE and just have
the Dockerfile extend the private image in its FROM clause.