Update README to mention middleman contentful
`data.aptible` is accessed in an number of places in helpers, but that
won't work if you haven't run `middleman contentful` first.

Ideally, we probably should either handle the absence of contentful data
gracefully or crash with an explicit error, but I don't really have the
familiarity with middleman / contentful to know how that might be done.