Remove Contentful runtime requirement
Instead of depending on the contentful_middleman gem to dynamically
generate content from Contentful resources at runtime, this PR treats
Contentful as a source of dumb static content, from which Markdown
content can be generated, which will live side-by-side with Markdown
files committed to the repo.

To do this, we introduce a `contentful:pull` Rake task, which uses the
more minimal contentful gem to download Contentful resources via the
Delivery and Preview APIs, and translate them into Markdown files that
reside in the same source/ directory as committed files.

All templates and layouts have also been updated to reflect the fact that
we no longer need Contentful at runtime:

- Middleman layout updates & cleanup for /resources and /blog
- Blog post layout: Get rid of blog/post.haml entirely
- Blog post layout: Remove layouts/blog_posts.haml (and move container
  div into blog/posts.haml, inheriting from the default layout)
- Blog post layout: Add the container div to /blog/author.haml and
- Inherit from default layout
- Resource layout: Get rid of resources/resource.haml entirely
- Resource layout: Update layouts/resource.haml to properly render
- Pagination specs
- Remove unused helper methods
- Remove references to `data.aptible.resource_pages`