Clean up the clean up
The old quickstart links would link to the framework if there was only
one. It makes more sense with the new docs to link to the language
page. Then we don’t need to worry about additions or changes to a
framework link. — thanks @krallin

Update the `BASE_DOC_URL` for staging so that www’s staging environment
will point to staging docs — thanks @krallin

Added a note to the README on when and how to use the `BASE_DOC_URL`
when running docs and www locally.