Allow links in the changelog index
Two changes here:

- Added a flag in Contentful (ran this by Gib) to indicate that a given
  Changelog entry's body (as opposed to its excerpt) should be rendered
  in the list view. This will let us use links there, since the
  Changelog entry's body is HTML.
- Removed the link surrounding each chngelog item in the index (because
  that prevents us from having links in it), and replaced them with a
  "Read More" link at the end.

Even though there is nothing more to read, I opted to retain the "Read
More" link for changelog posts that are displayed in full (i.e. those
that have the flag set) in the changelog index, for two reasons:

- Consistency.
- So users can access a URL for the post to share.

I'm totally to reverting that last decision, or changing the copy (e.g.
to say "to the post" instead of "read more").