Link resource cover images
Resources should include documents, a contentful object type, that
references PDFs that are committed to the repo.

If there is a single document, I assume it is represented by a cover
image and link the image to the document as a temporary solution to our
current Enclave Architecture Reference issues (see
https://trello.com/c/ngDQnkPp/137-resource-cover-image-improve-ux)

The image cover image can also be removed.